requirements.txt
arrow==0.10.0
cffi==1.9.1
cryptography==1.8.1
decorator==4.0.11
enum34==1.1.6
google-api-python-client==1.6.2
httplib2==0.10.3
idna==2.5
ipaddress==1.0.18
oauth2client==4.0.0
py==1.4.33
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycparser==2.17
PyDrive==1.3.1
pyOpenSSL==16.2.0
python-dateutil==2.6.0
python-slugify==1.2.1
pytz==2017.2
PyYAML==3.12
requests==2.18.1
retry==0.9.2
rsa==3.4.2
simplejson==3.10.0
six==1.10.0
Unidecode==0.4.19
uritemplate==3.0.0
pygsheets==1.1.1
